% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_rectangular.R
\name{distance_to_rectangular}
\alias{distance_to_rectangular}
\title{Convert pairwise distance matrices to rectangular matrices}
\usage{
distance_to_rectangular(
  distance.matrix,
  labels = names(distance.matrix),
  n.dimensions = 2,
  metric = TRUE,
  ...
)
}
\arguments{
\item{distance.matrix}{Distance matrix (i.e. object of class 'dist'). Can be created using the function \code{\link[stats]{dist}} or converted to using  \code{\link[stats]{as.dist}}.}

\item{labels}{Character vector or factor containing labels to be used for rows/columns in the output data frame. Default is \code{names(distance.matrix)}. Must be the same length as the number of observations in 'distance.matrix'.}

\item{n.dimensions}{Integer vector of length 1 indicating the number of of dimensions to represent distances in a new space. Default is 2.}

\item{metric}{Logical argument to control if Metric (a.k.a. Classical, \code{TRUE}, default) or Non-Metric MUltidimensional Scaling (\code{FALSE}) is used to project in a new n-dimension space. Non-Metric MDS is conducted using the function \code{\link[MASS]{isoMDS}} while Classical MDS uses the function \code{\link[stats]{cmdscale}}. So yes, it is a silly wrapper over those 2 functions.}

\item{...}{Additional arguments to be passed to \code{\link[caret]{train}} (only used for machine learning models).}
}
\value{
A data frame with the new dimensions representing the position of observations in a new n-dimension space. If \code{metric = FALSE} the output data frame is embedded in a list that also includes the stress value.
}
\description{
\code{distance_to_rectangular} converts binary triangular matrices to rectangular matrices using Multidimensional Scaling.
}
\details{
It is a silly wrapper over 2 multidimensional scaling functions (\code{\link[MASS]{isoMDS}} and \code{\link[stats]{cmdscale}}) that simplifies the calculation of Multidimensional Scaling and formatting of its output to be used with other functions in the package.
}
\examples{
{
data("example_space")

dist_example <- dist(example_space[example_space$group \%in\% c("G1", "G2"), 
c("dimension_1", "dimension_2")])

# convert into a 2-dimension space
rect_example <- distance_to_rectangular(distance.matrix = dist_example,
metric = TRUE)

head(rect_example)

\donttest{
# convert into a 2-dimension space with non-metric MDS
rect_example <- distance_to_rectangular(distance.matrix = dist_example, 
metric = FALSE, maxit = 3)
 }
}
}
\references{
Araya-Salas, M, & K. Odom. 2022, PhenotypeSpace: an R package to quantify and compare phenotypic trait spaces R package version 0.1.0.
}
\seealso{
\code{\link{distance_to_rectangular}}, \code{\link{rectangular_to_triangular}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
