% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_space-data.R
\docType{data}
\name{example_space}
\alias{example_space}
\title{Example bi-dimensional space data}
\format{
An object of class \code{data.frame} with 1550 rows and 4 columns.
}
\source{
Marcelo Araya Salas, PhenotypeSpace
}
\usage{
data(example_space)
}
\description{
\code{example_space} is a data frame with 1550 observations of a simulated phenotypic bi-dimensional space including 5 groups.
}
\keyword{datasets}
