% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefact_space_overlap.R
\name{rarefact_space_overlap}
\alias{rarefact_space_overlap}
\title{Calculates rarefacted space overlaps}
\usage{
rarefact_space_overlap(X, dimensions, group, n = NULL, replace = FALSE, 
seed = NULL, parallel = 1, pb = TRUE, iterations = 30, ...)
}
\arguments{
\item{X}{Data frame containing columns for the dimensions of the phenotypic space (numeric) and a categorical or factor column with group labels.}

\item{dimensions}{Character vector with the names of columns containing the dimensions of the phenotypic space.}

\item{group}{Character vector with the name of the column (character or factor) containing group labels.}

\item{n}{Integer vector of length 1 indicating the number of samples to be use for rarefaction (i.e. how many samples per group will be gather at each iteration). Default is the minimum sample size across groups.}

\item{replace}{Logical argument to control if sampling is done with replacement. Default is \code{FALSE}.}

\item{seed}{Integer vector of length 1 setting the seed (see \code{\link[base]{set.seed}}). If used results should be the same on different runs, so it makes them replicable.}

\item{parallel}{Integer vector of length 1. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{iterations}{Integer vector of length 1. Controls how the number of times the rarefaction routine is iterated. Default is 30.}
}
\value{

}
\description{
\code{rarefact_space_overlap}
}
\details{

}
\examples{
{
# load data
data("example_space")

# get proportion of space that overlaps (try with more iterations on your own data)
prop_overlaps <- rarefact_space_overlap(
X = example_space,
dimensions =  c("Dimension_1", "Dimension_2"),
group = "ID",
type = "proportional", 
iterations = 5)

# get minimum convex polygon overlap for each group (non-symmetric)
mcp_overlaps <- rarefact_space_overlap(
X = example_space,
dimensions =  c("Dimension_1", "Dimension_2"),
group = "ID",
iterations = 5)

# convert to non-symmetric triangular matrix
rectangular_to_triangular(mcp_overlaps, symmetric = FALSE)
}
}
\references{
{
Araya-Salas, M, & K. Odom. 2022, PhenotypeSpace: quantifying phenotypic trait spaces. R package version 0.1.0.
}
}
\seealso{
\code{\link{space_overlap}}, \code{\link{rectangular_to_triangular}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
