% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_overlap.R
\name{space_overlap}
\alias{space_overlap}
\title{Estimates pairwise overlaps of phenotype spaces}
\usage{
space_overlap(X, dimensions, group, parallel = 1, type = "mcp", 
pb = TRUE, outliers = 0.95, iterations = 30)
}
\arguments{
\item{X}{Data frame containing columns for the dimensions of the phenotypic space (numeric) and a categorical or factor column with group labels.}

\item{dimensions}{Character vector with the names of the columns containing the dimensions of the phenotypic space.}

\item{group}{Character vector with the name of the column (character or factor) containing group labels.}

\item{parallel}{Integer vector of length 1. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{type}{}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{iterations}{Integer vector of length 1. Controls how the number of times the rarefaction routine is iterated. Default is 30.}
}
\value{

}
\description{
\code{space_overlap}
}
\details{

}
\examples{
{
# load data
data("example_space")

# get proportion of space that overlaps 
prop_overlaps <- space_overlap(
X = example_space,
dimensions =  c("Dimension_1", "Dimension_2"),
group = "ID",
type = "proportional")

#' # get symmetric triangular matrix
rectangular_to_triangular(prop_overlaps)

# get minimum convex polygon overlap for each group (non-symmetric)
mcp_overlaps <- space_overlap(
X = example_space,
dimensions =  c("Dimension_1", "Dimension_2"),
group = "ID",
type = "mcp")

# convert to non-symmetric triangular matrix
rectangular_to_triangular(mcp_overlaps, symmetric = FALSE)
}
}
\references{
{
Araya-Salas, M, & K. Odom. 2022, PhenotypeSpace: quantifying phenotypic trait spaces. R package version 0.1.0.
}
}
\seealso{
\code{\link{rarefact_space_overlap}}, \code{\link{rectangular_to_triangular}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
