% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_size.R
\name{space_size}
\alias{space_size}
\title{Estimates the size of phenotypic spaces}
\usage{
space_size(
  X,
  dimensions = NULL,
  group,
  cores = 1,
  type = "mcp",
  pb = TRUE,
  outliers = 0.95,
  ...
)
}
\arguments{
\item{X}{Data frame containing columns for the dimensions of the phenotypic space (numeric) and a categorical or factor column with group labels.}

\item{dimensions}{Character vector with the names of columns containing the dimensions of the phenotypic space.}

\item{group}{Character vector with the name of the column (character or factor) containing group labels.}

\item{cores}{Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{type}{Character vector of length 1. Controls the type of metric to be used for quantifying space size. Three metrics are available:
\itemize{
 \item \code{mcp}: minimum convex polygon area using the function  \code{\link[adehabitatHR]{mcp}}. The minimum sample size (per group) must be 2 observations.
 \item \code{density}: kernel density area using the function \code{\link[adehabitatHR]{kernelUD}}. The minimum sample size (per group) must be 6 observations.
 \item \code{mst}: minimum spanning tree using the function \code{\link[vegan]{spantree}}. The minimum sample size (per group) must be 2 observations. This method is expected to be more robust to the influence of outliers. . Note that mst is not a actually measuring area but distance between observations. However, it still help to quantify the size of the sub-region in trait space.   
 }}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{outliers}{Numeric vector of length 1. A value between 0 and 1 controlling the proportion of outlier observations to be excluded. Outliers are determined as those farthest away from the sub-space centroid.}

\item{...}{Additional arguments to be passed to \code{\link[adehabitatHR]{kernelUD}} for kernel density estimation (when  \code{type = 'density'}.}
}
\value{
A data frame containing the phenotypic space size for each group.
}
\description{
\code{space_size}
}
\details{
The function quantifies the size of the phenotypic sub-spaces.
}
\examples{
{
# load data
data("example_space")

# plot data
xs <- tapply(example_space$dimension_1, example_space$group, mean)
ys <- tapply(example_space$dimension_2, example_space$group, mean)
plot(example_space[, c("dimension_1", "dimension_2")], 
col = example_space$color, pch = 20, cex = 1.8)
text(xs, ys, labels = names(xs), cex = 2.5)

# MCP spaces
space_size(
X = example_space,
dimensions =  c("dimension_1", "dimension_2"),
group = "group",
type = "mcp")

# MST 
space_size(
  X = example_space,
  dimensions =  c("dimension_1", "dimension_2"),
  group = "group",
  type = "mst")
}
}
\references{
{
Araya-Salas, M, & K. Odom. 2022, PhenotypeSpace: an R package to quantify and compare phenotypic trait spaces R package version 0.1.0.
}
}
\seealso{
\code{\link{rarefact_space_size}}, \code{\link{space_size_difference}}, \code{\link{rarefact_space_size_difference}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
