% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_space-data.R
\docType{data}
\name{example_space}
\alias{example_space}
\title{Example bi-dimensional space data}
\format{
An object of class \code{data.frame} with 1550 rows and 4 columns.
}
\source{
Marcelo Araya Salas, PhenotypeSpace
}
\usage{
data(example_space)
}
\description{
\code{example_space} is a data frame with 1550 observations of a simulated phenotypic bi-dimensional space including 5 groups.
}
\examples{
{
# load data
data("example_space")

#plot space 
xs <- tapply(example_space$Dimension_1, example_space$ID, mean)
ys <- tapply(example_space$Dimension_2, example_space$ID, mean)
plot(example_space[, c("Dimension_1", "Dimension_2")], 
   col = example_space$color, pch = 20, cex = 1.8)
text(xs, ys, labels = names(xs), cex = 2.5)
}

}
\keyword{datasets}
