% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeSpace-package.R
\docType{package}
\name{PhenotypeSpace}
\alias{PhenotypeSpace}
\title{PhenotypeSpace: Quantifying phenotypic spaces}
\description{
Facilitates quantifying phenotypic trait spaces and comparing spaces across groups.
}
\details{
The package offers functions for:
  \itemize{
  \item Estimate absolute values of phenotypic trait spaces
  \item Compare phenotypic trait spaces across groups
  \item Modify output formats to facilitate statistical analysis
  }

All functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency.

License: GPL (>= 2)
}
\author{
Marcelo Araya-Salas

  Maintainer: Marcelo Araya-Salas (\email{marcelo.araya@ucr.ac.cr})
}
