% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefact_space_size_difference.R
\name{rarefact_space_size_difference}
\alias{rarefact_space_size_difference}
\title{Calculates rarefacted space size differences}
\usage{
rarefact_space_size_difference(X, dimensions, group, n = NULL, replace = FALSE, 
seed = NULL, parallel = 1, pb = TRUE, iterations = 30, ...)
}
\arguments{
\item{X}{Data frame containing columns for the dimensions of the phenotypic space (numeric) and a categorical or factor column with group labels.}

\item{dimensions}{Character vector with the names of columns containing the dimensions of the phenotypic space.}

\item{group}{Character vector with the name of the column (character or factor) containing group labels.}

\item{n}{Integer vector of length 1 indicating the number of samples to be use for rarefaction (i.e. how many samples per group will be gather at each iteration). Default is the minimum sample size across groups.}

\item{replace}{Logical argument to control if sampling is done with replacement. Default is \code{FALSE}.}

\item{seed}{Integer vector of length 1 setting the seed (see \code{\link[base]{set.seed}}). If used results should be the same on different runs, so it makes them replicable.}

\item{parallel}{Integer vector of length 1. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{iterations}{Integer vector of length 1. Controls how the number of times the rarefaction routine is iterated. Default is 30.}

\item{...}{Additional arguments to be passed to \code{\link{space_size_difference}}.}
}
\value{
A data frame containing the mean, minimum, maximum and standard deviation of the space size difference across iterations for each pair of groups.
}
\description{
\code{rarefact_space_size_difference}
}
\details{
The function applies a rarefaction sub-sampling procedure for evaluating pairwise space size differences (internally using \code{\link{space_size_difference}}). The size of a phenotypic space might change as a function of number of samples. Hence, ideally, spaces should be compared between groups of similar sample sizes. Rarefaction allows to compare groups of unbalanced sample sizes by randomly re-sampling observations using the same number samples across groups iteratively.
}
\examples{
{
# load data
data("example_space")

# get rarefied size difference using MCP (try with more iterations on your own data)
mcp_size_diff <- rarefact_space_size_difference(
X = example_space,
dimensions =  c("Dimension_1", "Dimension_2"),
group = "ID",
type = "mcp", 
iterations = 5)

# convert to non-symmetric triangular matrix
rectangular_to_triangular(mcp_size_diff, symmetric = FALSE)
}
}
\references{
{
Araya-Salas, M, & K. Odom. 2022, PhenotypeSpace: an R package to quantify and compare phenotypic trait spaces R package version 0.1.0.
}
}
\seealso{
\code{\link{rarefact_space_similarity}}, \code{\link{space_size_difference}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
