% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_size.R
\name{space_size}
\alias{space_size}
\title{Estimates the size of phenotypic spaces}
\usage{
space_size(X, dimensions = NULL, group, parallel = 1, type = "mcp", 
pb = TRUE, outliers = 0.95, proportional = FALSE, ...)
}
\arguments{
\item{X}{Data frame containing columns for the dimensions of the phenotypic space (numeric) and a categorical or factor column with group labels.}

\item{dimensions}{Character vector with the names of columns containing the dimensions of the phenotypic space.}

\item{group}{Character vector with the name of the column (character or factor) containing group labels.}

\item{parallel}{Integer vector of length 1. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).}

\item{pb}{Logical argument to control if progress bar is shown. Default is \code{TRUE}.}

\item{iterations}{Integer vector of length 1. Controls how the number of times the rarefaction routine is iterated. Default is 30.}
}
\value{

}
\description{
\code{space_size}
}
\details{

}
\examples{
{
# load data
data("example_space")

# plot data
xs <- tapply(example_space$Dimension_1, example_space$ID, mean)
ys <- tapply(example_space$Dimension_2, example_space$ID, mean)
plot(example_space[, c("Dimension_1", "Dimension_2")], 
col = example_space$color, pch = 20, cex = 1.8)
text(xs, ys, labels = names(xs), cex = 2.5)

# MCP spaces
space_size(
X = example_space,
dimensions =  c("Dimension_1", "Dimension_2"),
group = "ID",
type = "mcp")

# MST 
space_size(
  X = example_space,
  dimensions =  c("Dimension_1", "Dimension_2"),
  group = "ID",
  type = "mst")
}
}
\references{
{
Araya-Salas, M, & K. Odom. 2022, PhenotypeSpace: quantifying phenotypic trait spaces. R package version 0.1.0.
}
}
\seealso{
\code{\link{rarefact_space_size}}, \code{\link{space_size_difference}}, \code{\link{rarefact_space_size_difference}}
}
\author{
Marcelo Araya-Salas \email{marcelo.araya@ucr.ac.cr})
}
